pyproject.toml
[project]
name = "modernisation-platform-ui"
version = "1.0.0"
description = "Passionate engineers delivering great services"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "alembic==1.17.0",
  "authlib==1.6.6",
  "flask==3.1.3",
  "flask-cors==6.0.1",
  "flask-limiter==4.0.0",
  "flask-sqlalchemy==3.1.1",
  "govuk-frontend-jinja==3.8.0",
  "gunicorn==23.0.0",
  "psycopg2-binary==2.9.11",
  "pygithub==2.8.1",
  "pynacl==1.6.2",
  "sentry-sdk==2.41.0",
  "urllib3==2.6.3",
  "werkzeug==3.1.6",
]

[dependency-groups]
dev = [
  "coverage==7.10.7",
  "pre-commit==4.3.0",
  "pytest==8.4.2",
  "ruff==0.14.0",
  "zizmor==1.14.2",
]

[tool.sqlfluff.core]
dialect = "postgres"

# [tool.ruff.lint]
# select = [
#     # flake8-bugbear
#     "B",
#     # pycodestyle
#     "E",
#     # Pyflakes
#     "F",
#     # isort
#     "I",
#     # flake8-simplify
#     "SIM",
#     # pyupgrade
#     "UP",
# ]
